initial commit
this is a very simple python script that opens a hard-coded FreeCAD file (.FCStd), takes a screenshot of the model, and saves it as a PNG.

The intention is to use this as part of a Continuous Integration action, to automate updates to documentation by updating the doc's images of the CAD file.

For now, I'm using assembly.FCStd, which is a really great example of object that was designed as an example object for rendering-out to images with POV Ray Tracing. Attribution = schupin & vocx

 * https://forum.freecad.org/memberlist.php?mode=viewprofile&u=18259
 * https://forum.freecad.org/memberlist.php?mode=viewprofile&u=21943
 * https://forum.freecad.org/viewtopic.php?f=36&t=32745
 * https://wiki.freecad.org/Tutorial_FreeCAD_POV_ray

TODO: Actually make a GitHub Actions for this